requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
requests>=2.28.0
openai>=1.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0